finalized project

+++ src/main/webapp/viewCars.jsp
@@ -12,7 +12,31 @@
     <!-- Custom CSS -->
     <style>
         body {
-            background-color: #f8f9fa;
+            background: url('https://www.shutterstock.com/image-vector/taxi-service-abstract-banner-design-600nw-2506919915.jpg') no-repeat center center fixed;
+            background-size: cover;
+            position: relative;
+            display: flex;
+            flex-direction: column;
+            min-height: 100vh;
+        }
+        body::before {
+            content: '';
+            position: absolute;
+            top: 0;
+            left: 0;
+            width: 100%;
+            height: 100%;
+            background-color: rgba(0, 0, 0, 0.5);
+            z-index: -1;
+        }
+        .navbar {
+            background-color: #ff8800; 
+        }
+        .navbar-brand, .nav-link {
+            color: #fff !important;
+        }
+        .nav-link:hover {
+            color: #f8f9fa !important;
         }
         .container {
             margin-top: 2rem;
@@ -34,6 +58,7 @@
     </style>
 </head>
 <body>
+        <!-- Navbar -->
     <!-- Navbar -->
     <nav class="navbar navbar-expand-lg navbar-dark">
         <div class="container-fluid">
@@ -43,6 +68,16 @@
             </button>
             <div class="collapse navbar-collapse" id="navbarNav">
                 <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+                    <!-- Manage Users Dropdown -->
+                    <li class="nav-item dropdown">
+                        <a class="nav-link dropdown-toggle" href="#" id="navbarDropdownUsers" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                            Manage Users
+                        </a>
+                        <ul class="dropdown-menu" aria-labelledby="navbarDropdownUsers">
+                            <li><a class="dropdown-item" href="${pageContext.request.contextPath}/addUser.jsp">Add User</a></li>
+                            <li><a class="dropdown-item" href="${pageContext.request.contextPath}/users">View Users</a></li>
+                        </ul>
+                    </li>
                     <!-- Manage Customers Dropdown -->
                     <li class="nav-item dropdown">
                         <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
@@ -69,16 +104,18 @@
                             Manage Cars
                         </a>
                         <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-                            <li><a class="dropdown-item" href="${pageContext.request.contextPath}/addCar.jsp">Add Car</a></li>
+                            <li><a class="dropdown-item" href="${pageContext.request.contextPath}/cars?action=add">Add Car</a></li>
                             <li><a class="dropdown-item" href="${pageContext.request.contextPath}/cars">View Cars</a></li>
                         </ul>
                     </li>
+                    <!-- Manage Bookings Dropdown -->
                     <li class="nav-item">
                         <a class="nav-link" href="${pageContext.request.contextPath}/bookings">Manage Bookings</a>
                     </li>
+                    <!-- Manage Bills  -->
                     <li class="nav-item">
-                        <a class="nav-link" href="${pageContext.request.contextPath}/bills">Manage Bills</a>
-                    </li>
+						    <a class="nav-link" href="${pageContext.request.contextPath}/bills">Manage Bills</a>
+						</li>
                 </ul>
                 <form action="${pageContext.request.contextPath}/login" method="get" class="d-flex">
                     <button type="submit" class="btn btn-outline-light">Logout</button>
@@ -92,6 +129,7 @@
             <div class="col-md-10">
                 <div class="table-container">
                     <h2>Car List</h2>
+                    <a href="${pageContext.request.contextPath}/cars?action=add" class="btn btn-primary mb-3">Add Car</a>
                     <table class="table table-striped">
                         <thead>
                             <tr>
@@ -112,7 +150,7 @@
                                     <td><%= car.getDriverId() %></td>
                                     <td>
                                         <a href="${pageContext.request.contextPath}/cars?action=edit&id=<%= car.getCarId() %>" class="btn btn-primary btn-action">Edit</a>
-                                        <a href="${pageContext.request.contextPath}/cars?action=delete&id=<%= car.getCarId() %>" class="btn btn-danger btn-action">Delete</a>
+                                        <a href="${pageContext.request.contextPath}/cars?action=delete&id=<%= car.getCarId() %>" class="btn btn-danger btn-action" onclick="return confirm('Are you sure you want to delete this car?');">Delete</a>
                                     </td>
                                 </tr>
                             <% } %>
