+++ src/main/webapp/viewDrivers.jsp
@@ -11,11 +11,36 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
     <!-- Custom CSS -->
     <style>
-        body {
-            background-color: #f8f9fa;
-        }
+
         .container {
             margin-top: 2rem;
+        }
+        body {
+            background: url('https://www.shutterstock.com/image-vector/taxi-service-abstract-banner-design-600nw-2506919915.jpg') no-repeat center center fixed;
+            background-size: cover;
+            position: relative;
+            display: flex;
+            flex-direction: column;
+            min-height: 100vh;
+        }
+        body::before {
+            content: '';
+            position: absolute;
+            top: 0;
+            left: 0;
+            width: 100%;
+            height: 100%;
+            background-color: rgba(0, 0, 0, 0.5);
+            z-index: -1;
+        }
+        .navbar {
+            background-color: #ff8800; 
+        }
+        .navbar-brand, .nav-link {
+            color: #fff !important;
+        }
+        .nav-link:hover {
+            color: #f8f9fa !important;
         }
         .table-container {
             background: #fff;
@@ -34,7 +59,8 @@
     </style>
 </head>
 <body>
-    <!-- Navbar -->
+        <!-- Navbar -->
+	    <!-- Navbar -->
     <nav class="navbar navbar-expand-lg navbar-dark">
         <div class="container-fluid">
             <a class="navbar-brand" href="${pageContext.request.contextPath}/dashboard.jsp">Mega City Cab</a>
@@ -43,6 +69,16 @@
             </button>
             <div class="collapse navbar-collapse" id="navbarNav">
                 <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+                    <!-- Manage Users Dropdown -->
+                    <li class="nav-item dropdown">
+                        <a class="nav-link dropdown-toggle" href="#" id="navbarDropdownUsers" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                            Manage Users
+                        </a>
+                        <ul class="dropdown-menu" aria-labelledby="navbarDropdownUsers">
+                            <li><a class="dropdown-item" href="${pageContext.request.contextPath}/addUser.jsp">Add User</a></li>
+                            <li><a class="dropdown-item" href="${pageContext.request.contextPath}/users">View Users</a></li>
+                        </ul>
+                    </li>
                     <!-- Manage Customers Dropdown -->
                     <li class="nav-item dropdown">
                         <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
@@ -63,15 +99,24 @@
                             <li><a class="dropdown-item" href="${pageContext.request.contextPath}/drivers">View Drivers</a></li>
                         </ul>
                     </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="${pageContext.request.contextPath}/cars">Manage Cars</a>
+                    <!-- Manage Cars Dropdown -->
+                    <li class="nav-item dropdown">
+                        <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                            Manage Cars
+                        </a>
+                        <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
+                            <li><a class="dropdown-item" href="${pageContext.request.contextPath}/cars?action=add">Add Car</a></li>
+                            <li><a class="dropdown-item" href="${pageContext.request.contextPath}/cars">View Cars</a></li>
+                        </ul>
                     </li>
+                    <!-- Manage Bookings Dropdown -->
                     <li class="nav-item">
                         <a class="nav-link" href="${pageContext.request.contextPath}/bookings">Manage Bookings</a>
                     </li>
+                    <!-- Manage Bills  -->
                     <li class="nav-item">
-                        <a class="nav-link" href="${pageContext.request.contextPath}/bills">Manage Bills</a>
-                    </li>
+						    <a class="nav-link" href="${pageContext.request.contextPath}/bills">Manage Bills</a>
+						</li>
                 </ul>
                 <form action="${pageContext.request.contextPath}/login" method="get" class="d-flex">
                     <button type="submit" class="btn btn-outline-light">Logout</button>
@@ -85,6 +130,7 @@
             <div class="col-md-10">
                 <div class="table-container">
                     <h2>Driver List</h2>
+                    <a href="${pageContext.request.contextPath}/addDriver.jsp" class="btn btn-primary mb-3">Add Driver</a>
                     <table class="table table-striped">
                         <thead>
                             <tr>
@@ -105,7 +151,7 @@
                                     <td><%= driver.getContact() %></td>
                                     <td>
                                         <a href="${pageContext.request.contextPath}/drivers?action=edit&id=<%= driver.getDriverId() %>" class="btn btn-primary btn-action">Edit</a>
-                                        <a href="${pageContext.request.contextPath}/drivers?action=delete&id=<%= driver.getDriverId() %>" class="btn btn-danger btn-action">Delete</a>
+                                        <a href="${pageContext.request.contextPath}/drivers?action=delete&id=<%= driver.getDriverId() %>" class="btn btn-danger btn-action" onclick="return confirm('Are you sure you want to delete this driver?');">Delete</a>
                                     </td>
                                 </tr>
                             <% } %>
